move rpc client to rpc module
