--- pkg/client/starknet_client_test.go
@@ -1,2 +1,0 @@
-package client
-
