+++ pkg/rpc/starknet_client.go
@@ -1,4 +1,4 @@
-package client
+package rpc
 
 import (
 	"context"
